pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Gene Expression Analysis"
authors = ["Karim Wahba <karim.ibrahim.amin@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^2.0"
pyyaml = "^6.0.2"
logging = "^0.4.9.6"
pathlib = "^1.0.1"
matplotlib = "3.4.3"
scikit-learn = "0.24.2"
seaborn = "0.11.1"

[[tool.poetry.packages]]
include = "utils"
from = "src"

[[tool.poetry.packages]]
include = "configs"
from = "src"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "Gene_Expression_Analysis"
version = "0.1.0"
description = "Effect size vs Sample Size for Gene Expression"
authors = [{name = "Karim Wahba and Jenny Arndt", email = "karim.ibrahim.amin@gmail.com"}]